Prevent new rules from resulting in silent deletions of third-party files
Some third-party files must be declared as dependencies in order to ensure that
their submodules are checked out; however, before this resulted in their
automatic deletion as intermediates unless the author of the rule remembered to
explicitly declare them as secondaries. This fixes the situation by always
implicitly considering them precious, removing the treacherous extra step.